concept_gcp_sre/automation/requirements.txt
google-cloud-run
google-cloud-monitoring
slack_sdk
